consolidates IterableInboxMessageListItem component into IterableInboxMessageCell component
